Moved button functionality to relate it to chatbots
Preparing for future refactors
