+++ buttons.py
@@ -53,11 +53,11 @@ class HVZButton(discord.ui.Button):
             custom_id = function.__name__
         else:
             custom_id = str(custom_id)
-        if label is not None:
+        if label:
             label = str(label)
         else:
             label=config['buttons'][function.__name__]['label']
-        if style is None:
+        if not style:
             if color is None:
                 color = config['buttons'][function.__name__]['color']
 
@@ -89,53 +89,5 @@ class HVZButtonCog(commands.Cog):
     """A cog with a slash command for posting the message with buttons
     and to initialize the view again when the bot is restarted
     """
-    button_options = []
-    for option in config['buttons']:
-        button_options.append(option)
+    pass
 
-    def __init__(self, bot):
-        self.bot = bot
-        button_options = []
-        for option in config['buttons']:
-            button_options.append(option)
-
-        # make sure to set the guild ID here to whatever server you want the buttons in
-        @bot.command(guild_ids=guild_id_list)
-        async def post(
-                ctx,
-                function_name: Option(str, 'Which button to post.', choices=button_options, name='function',
-                                      required=True),
-                message: Option(str, 'Message to replace the default.', required=False, default=None)
-        ):
-            """Post a message with a button: registration or tag log. Can change message."""
-
-            # timeout is None because we want this view to be persistent
-            view = discord.ui.View(timeout=None)
-            function = getattr(self.bot, function_name)
-            view.add_item(HVZButton(function, custom_id=function_name))
-
-            if message is None:
-                message = config['buttons'][function_name]['message']
-
-            await ctx.channel.send(message, view=view)
-            await ctx.respond('Posted message with button.', ephemeral=True)
-
-    @commands.Cog.listener()
-    async def on_ready(self):
-        """This function is called every time the bot restarts.
-        If a view was already created before (with the same custom IDs for buttons)
-        it will be loaded and the bot will start watching for button clicks again.
-        """
-
-        # we recreate the view as we did in the /post command
-        view = discord.ui.View(timeout=None)
-        # make sure to set the guild ID here to whatever server you want the buttons in
-        # guild = self.bot.guild
-        # loop through the list of roles and add a new button to the view for each role
-        for function_name in config['buttons']:
-            # get the function to call from bot based on the name in config
-            function = getattr(self.bot, function_name)
-            view.add_item(HVZButton(function, custom_id=function_name))
-
-        # add the view to the bot so it will watch for button interactions
-        self.bot.add_view(view)
